[VFABI] Move the Vector ABI demangling utility to LLVMCore.
This fixes #71892 allowing us to check magled names in the IR verifier.
