+++ llvm/lib/IR/CMakeLists.txt
@@ -69,6 +69,7 @@ add_llvm_component_library(LLVMCore
   ValueSymbolTable.cpp
   VectorBuilder.cpp
   Verifier.cpp
+  VFABIDemangler.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LLVM_MAIN_INCLUDE_DIR}/llvm/IR
