+++ llvm/utils/gn/secondary/llvm/unittests/Analysis/BUILD.gn
@@ -50,7 +50,6 @@ unittest("AnalysisTests") {
     "UnrollAnalyzerTest.cpp",
     "ValueLatticeTest.cpp",
     "ValueTrackingTest.cpp",
-    "VectorFunctionABITest.cpp",
     "VectorUtilsTest.cpp",
   ]
