+++ llvm/utils/gn/secondary/llvm/unittests/IR/BUILD.gn
@@ -52,5 +52,6 @@ unittest("IRTests") {
     "VectorBuilderTest.cpp",
     "VectorTypesTest.cpp",
     "VerifierTest.cpp",
+    "VFABIDemanglerTest.cpp,"
   ]
 }
